Swap `modBy` argument order to clarify logic
When going through the koans, I noticed that the anonymous functions
using `modBy` were producing results that could be clearer. While I
don't know the intention of the author for certain, it felt like the
functions were not working as intended.

By swapping argument order, I found the functions to be more intuitive.
Perhaps others will as well :)

For example:
```elm
~/oss/elm-koans $ elm repl
> unsortedList = [ 10, 6, 8 ]
[10,6,8] : List number
> List.sortBy (\x -> modBy x 3) unsortedList
[10,6,8] : List Int
> List.sortBy (\x -> modBy x 3) unsortedList == unsortedList
True

-- with updated argument order
> List.sortBy (\x -> modBy 3 x) unsortedList
[6,10,8] : List Int
```